Versions Details
This game is not made responsive. So, specific app for specific resolution and platform is provided. Although this game contains only one level but, the game is really attractive and indulging. The android version has issues in the storing data so it will show your current score as the highscore.